* net mon fix for non-ta items with sub-members -- need base offset * #SHOW on #GUI_READ_ONLY items in state

+++ src/emergent/leabra/LeabraLayerSpec_core.h
@@ -652,7 +652,7 @@
       float& lvar = net->ThrLayStats(thr_no, li, 0, LEABRA_NETWORK_STATE::NETSD);
       var += lvar;
     }
-    lay->net_sd = sqrt(var);
+    lay->net_sd = sqrtf(var);
     lay->avg_net_sd.Increment(lay->net_sd);
     return var;
   }
